Refactor fuzzy match threshold to use double precision
Updated fuzzy match threshold logic to use `double` instead of `float` for improved precision and consistency. Adjusted the default `FuzzyMatchThreshold` value from `0.5` to `0.7` and updated test cases to reflect the new threshold (`0.8`).

Revised calculations to round to 10 significant decimal places before flooring, ensuring consistent behavior and avoiding floating-point artifacts. Updated the `AllowedEdits` method, `FuzzyMatch` class, and `GrepEngineInitParams` to align with the new `double`-based threshold.

+++ dnGREP.Engines/GrepEngineBase.cs
@@ -977,7 +977,7 @@ namespace dnGREP.Engines
         {
             LinesBefore = 0;
             LinesAfter = 0;
-            FuzzyMatchThreshold = 0.7f;
+            FuzzyMatchThreshold = 0.7;
             VerboseMatchCount = false;
             // keep the default false for unit tests 
             SearchParallel = false;
@@ -987,14 +987,14 @@ namespace dnGREP.Engines
         {
             LinesBefore = linesBefore;
             LinesAfter = linesAfter;
-            FuzzyMatchThreshold = (float)fuzzyMatchThreshold;
+            FuzzyMatchThreshold = fuzzyMatchThreshold;
             VerboseMatchCount = verboseMatchCount;
             SearchParallel = searchParallel;
         }
 
         public int LinesBefore { get; private set; }
         public int LinesAfter { get; private set; }
-        public float FuzzyMatchThreshold { get; private set; }
+        public double FuzzyMatchThreshold { get; private set; }
         public bool VerboseMatchCount { get; private set; }
         public bool SearchParallel { get; private set; }
     }
